given up for now
When you submit the sites there's no way to update the info in utilities
(or at least not without a very bizarre workaround). I also need to look
into tab focus and things like that because switching from ne tab to the
other is still not accounted for in tracking.